refactor: 게이트웨이 util 분리
